+++ examples/mursst-icechunk-with-lithops/main.py
@@ -5,7 +5,7 @@ Main entry point for the lithops package.
 This script provides a simple interface for running the package from the command line.
 """
 
-from lithops_package.cli import main
+from cli import main
 
 if __name__ == "__main__":
     main()
