+++ examples/mursst-icechunk-with-lithops/models.py
